// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const EDUCATION_LEVEL = {
  HIGH_SCHOOL: 'HIGH_SCHOOL',
  COLLAGE: 'COLLAGE',
  BACHELORS: 'BACHELORS',
  MASTERS: 'MASTERS',
  PhD: 'PhD',
  OTHER: 'OTHER'
} as const

export type EDUCATION_LEVEL = (typeof EDUCATION_LEVEL)[keyof typeof EDUCATION_LEVEL]


export const INTEREST = {
  SCIENCE: 'SCIENCE',
  TECHNOLOGY: 'TECHNOLOGY',
  HISTORY: 'HISTORY',
  LITERATURE: 'LITERATURE',
  MATHEMATICS: 'MATHEMATICS',
  GEOGRAPHY: 'GEOGRAPHY',
  SPORTS: 'SPORTS',
  ENTERTAINMENT: 'ENTERTAINMENT',
  POLITICS: 'POLITICS',
  ART: 'ART',
  MUSIC: 'MUSIC',
  PHILOSOPHY: 'PHILOSOPHY',
  OTHER: 'OTHER'
} as const

export type INTEREST = (typeof INTEREST)[keyof typeof INTEREST]